Hahahaha!
If you read this, you probably followed the backports of the
clean-msg-after-fixup-continue backports (Junio decided that he likes a
different name better, he uses
js/rebase-i-clean-msg-after-fixup-continue instead).

Well, the problem here is not only that there were *tons* of merge
conflicts to resolve on the backport journey so far, but with this
version, we also had to resolve issues where flags were passed to
run_git_commit() as individual parameters instead of an `unsigned
flags`.

Oh, and the rebase--helper is not even used in v2.12.5 when running `git
rebase --skip`. Instead, the Unix shell version of `rebase -i` does all
of that, and I am really not willing to spend that much time on
backports to Unix shell scripts, fixing bugs that nobody even bothered
to report before I dug into them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>