Let the code compile with GCC 8
As with every major GCC version, we have to adjust our code to compile
with it.

In this instance, GCC acquired a distaste for casting function pointers
between different signatures. The common workaround is simply to cast to
`(void *)` first, and only then to the target signature. So let's just
do that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>